requirements.txt
asgiref==3.5.2
Django==3.0.5
django-ckeditor==6.5.1
django-environ==0.9.0
django-js-asset==2.0.0
gunicorn==20.1.0
mysqlclient==2.1.1
Pillow==9.3.0
pytz==2022.6
sqlparse==0.4.3
tzdata==2022.7
